fix(HNG-202): fix merge issues on dashboard
